Network stack lifecycle management
It turns out that we need to cater for upstream users wanting to drop
the currently-running network stack, with the expectation of all
resources being released. Due to the hilarity that is async Rust, we
thus must make sure all spawned tasks finish when the last reference to
the protocol is dropped, which indeed requires manual refcounting. Also
note that protocol handles registered as `libgit2` transports are now
tracked as weak pointers.